--- src/main/java/com/bookbla/americano/domain/alarm/controller/dto/request/MemberTokenCreateRequest.java
@@ -1,14 +1,0 @@
-package com.bookbla.americano.domain.alarm.controller.dto.request;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-@AllArgsConstructor
-public class MemberTokenCreateRequest {
-
-    private String token;
-}
