--- src/main/java/com/bookbla/americano/domain/alarm/service/AlarmService.java
@@ -1,16 +1,0 @@
-package com.bookbla.americano.domain.alarm.service;
-
-import com.bookbla.americano.domain.alarm.controller.dto.request.PushAlarmAllCreateRequest;
-import com.bookbla.americano.domain.alarm.controller.dto.request.PushAlarmCreateRequest;
-import com.bookbla.americano.domain.alarm.controller.dto.response.PushAlarmAllCreateResponse;
-import com.bookbla.americano.domain.alarm.controller.dto.response.PushAlarmCreateResponse;
-import com.bookbla.americano.domain.member.repository.entity.Member;
-
-public interface AlarmService {
-
-    PushAlarmCreateResponse sendPushAlarm(PushAlarmCreateRequest pushAlarmCreateRequest);
-
-    void sendPushAlarm(Member member, String title, String body);
-
-    PushAlarmAllCreateResponse sendPushAlarmAll(PushAlarmAllCreateRequest pushAlarmAllCreateRequest);
-}
