--- src/main/java/com/bookbla/americano/domain/alarm/service/dto/AlarmResponse.java
@@ -1,5 +1,0 @@
-package com.bookbla.americano.domain.alarm.service.dto;
-
-public interface AlarmResponse {
-
-}
