--- src/main/java/com/bookbla/americano/domain/alarm/service/dto/AlarmSendResponse.java
@@ -1,15 +1,0 @@
-package com.bookbla.americano.domain.alarm.service.dto;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Getter
-public class AlarmSendResponse {
-
-    private final String success;
-    private final String title;
-    private final String body;
-    private final String token;
-
-}
