+++ src/main/java/com/bookbla/americano/domain/notification/service/dto/NotificationResponse.java
@@ -1,0 +1,5 @@
+package com.bookbla.americano.domain.notification.service.dto;
+
+public interface NotificationResponse {
+
+}
